Split UI and logic into separate modules
